requirements.txt
absl-py==2.1.0
aenum==3.1.15
affine==2.4.0
aiohttp==3.9.5
aiosignal==1.3.1
albucore==0.0.12
albumentations==1.4.10
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
anyio==4.4.0
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asttokens==2.4.1
async-lru==2.0.4
async-timeout==4.0.3
attrdict==2.0.1
attrs==23.2.0
Babel==2.15.0
backports.tarfile==1.2.0
beautifulsoup4==4.12.3
bleach==6.1.0
blobfile==3.0.0
braceexpand==0.1.7
cachetools==5.3.3
certifi==2024.6.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
click-plugins==1.1.1
cligj==0.7.2
cmake==3.20.3
comm==0.2.2
contourpy==1.2.1
cryptography==42.0.8
cycler==0.12.1
dataclasses-json==0.6.7
dctorch==0.1.2
debugpy==1.8.1
decorator==5.1.1
defusedxml==0.7.1
distlib==0.3.8
docker-pycreds==0.4.0
dominate==2.9.1
einops==0.8.0
exceptiongroup==1.2.1
executing==2.0.1
fastjsonschema==2.20.0
filelock==3.15.4
flash_attn==2.5.9.post1
fonttools==4.53.0
fqdn==1.5.1
frozenlist==1.4.1
fsspec==2024.6.0
ftfy==6.2.0
fvcore==0.1.5.post20221221
gitdb==4.0.11
GitPython==3.1.43
grpcio==1.67.1
h11==0.14.0
hatch==1.12.0
hatchling==1.25.0
httpcore==1.0.5
httpx==0.27.0
huggingface-hub==0.23.4
hyperlink==21.0.0
idna==3.7
imageio==2.34.2
importlib_metadata==8.0.0
iopath==0.1.10
ipykernel==6.29.4
ipython==8.25.0
ipywidgets==8.1.3
isoduration==20.11.0
jaraco.classes==3.4.0
jaraco.context==5.3.0
jaraco.functools==4.0.1
jedi==0.19.1
jeepney==0.8.0
Jinja2==3.1.4
joblib==1.4.2
json5==0.9.25
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.22.0
jsonschema-specifications==2023.12.1
jupyter==1.0.0
jupyter-console==6.6.3
jupyter-events==0.10.0
jupyter-lsp==2.2.5
jupyter_client==8.6.2
jupyter_core==5.7.2
jupyter_server==2.14.1
jupyter_server_terminals==0.5.3
jupyterlab==4.2.2
jupyterlab_pygments==0.3.0
jupyterlab_server==2.27.2
jupyterlab_widgets==3.0.11
keyring==25.2.1
kiwisolver==1.4.5
kornia==0.7.2
kornia_rs==0.1.3
lazy_loader==0.4
lightgbm==4.4.0
lightning==2.3.0
lightning-utilities==0.11.2
lpips==0.1.4
lxml==5.3.0
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==2.1.5
marshmallow==3.21.3
matplotlib==3.9.2
matplotlib-inline==0.1.7
mdurl==0.1.2
mistune==3.0.2
more-itertools==10.3.0
mpmath==1.3.0
multidict==6.0.5
mypy-extensions==1.0.0
natsort==8.4.0
# natten==0.17.1+torch220cu118
nbclient==0.10.0
nbconvert==7.16.4
nbformat==5.10.4
nest-asyncio==1.6.0
networkx==3.3
notebook==7.2.1
notebook_shim==0.2.4
numpy==1.26.4


nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-ml-py==12.535.161
nvidia-nccl-cu12==2.19.3
nvidia-nvjitlink-cu12==12.5.82
nvidia-nvtx-cu12==12.1.105

# nvidia-cublas-cu11==11.11.3.6 
# nvidia-cuda-cupti-cu11==11.8.87 
# nvidia-cuda-nvrtc-cu11==11.8.89 
# nvidia-cuda-runtime-cu11==11.8.89 
# nvidia-cudnn-cu11==8.6.0.163 
# nvidia-cufft-cu11==10.9.0.58 
# nvidia-curand-cu11==10.3.0.86 
# nvidia-cusolver-cu11==11.4.1.48 
# nvidia-cusparse-cu11==11.7.5.86 
# nvidia-nccl-cu11==2.16.5 
# nvidia-nvtx-cu11==11.8.86

nvitop==1.3.2
oauthlib==3.2.2
omegaconf==2.3.0
open-clip-torch==2.24.0
opencv-python-headless==4.10.0.84
overrides==7.7.0
packaging==24.1
pandas==2.2.3
pandocfilters==1.5.1
parso==0.8.4
pathspec==0.12.1
pexpect==4.9.0
pillow==10.4.0
platformdirs==4.2.2
pluggy==1.5.0
portalocker==2.10.1
prometheus_client==0.20.0
prompt_toolkit==3.0.47
protobuf==5.28.0
psutil==6.0.0
ptyprocess==0.7.0
pure_eval==0.2.3
pycparser==2.22
pycryptodomex==3.21.0
pydantic==2.8.2
pydantic_core==2.20.1
Pygments==2.18.0
pyparsing==3.1.4
pyproj==3.6.1
python-dateutil==2.9.0.post0
python-json-logger==2.0.7
pytorch-fid==0.3.0
pytorch-lightning==2.3.3
pytz==2024.2
PyYAML==6.0.1
pyzmq==26.0.3
qtconsole==5.5.2
QtPy==2.4.1
rasterio==1.3.11
referencing==0.35.1
regex==2024.7.24
requests==2.32.3
requests-oauthlib==2.0.0
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==13.7.1
rpds-py==0.19.1
s2cloudless==1.7.2
safetensors==0.4.4
scikit-image==0.24.0
scikit-learn==1.5.1
scipy==1.14.0
SecretStorage==3.3.3
Send2Trash==1.8.3
sentencepiece==0.2.0
sentinelhub==3.10.3
sentry-sdk==2.13.0
setproctitle==1.3.3
shapely==2.0.6
shellingham==1.5.4
six==1.16.0
smmap==5.0.1
sniffio==1.3.1
snuggs==1.4.7
soupsieve==2.5
stack-data==0.6.3
sympy==1.13.1
tabulate==0.9.0
tensorboard==2.18.0
tensorboard-data-server==0.7.2
termcolor==2.4.0
terminado==0.18.1
threadpoolctl==3.5.0
tifffile==2024.7.24
timm==1.0.9
tinycss2==1.3.0
tokenizers==0.19.1
tomli==2.0.1
tomli_w==1.0.0
tomlkit==0.13.0
# torch==2.2.0+cu118
# torchaudio==2.2.0+cu118
torchdata==0.7.1
torchmetrics==1.4.0.post0
torchnet==0.0.4
# torchvision==0.17.0+cu121
tornado==6.4.1
tqdm==4.66.4
traitlets==5.14.3
transformers==4.44.2
triton==2.2.0
trove-classifiers==2024.7.2
types-python-dateutil==2.9.0.20240316
typing-inspect==0.9.0
typing_extensions==4.12.2
tzdata==2024.2
uri-template==1.3.0
urllib3==2.2.2
userpath==1.9.2
utm==0.7.0
uv==0.2.29
virtualenv==20.26.3
visdom==0.2.4
wandb==0.17.8
wcwidth==0.2.13
webcolors==24.6.0
webdataset==0.2.96
webencodings==0.5.1
websocket-client==1.8.0
Werkzeug==3.0.6
widgetsnbextension==4.0.11
xformers==0.0.25
yacs==0.1.8
yarl==1.9.4
zipp==3.19.2
zstandard==0.23.0
